universeg_aitrox/requirements.txt
einops
torch
pydantic
numpy
Pillow
nibabel
pandas
SimpleITK
pydicom
opencv-python
